# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "JavInfoApi",
    "description": "媒体元数据查询 API，基于 r18.dev 数据库（PostgreSQL）",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "本地开发服务器"
    }
  ],
  "paths": {
...
    "/api/v1/actresses": {
      "get": {
        "summary": "演员列表",
        "description": "获取演员分页列表",
        "operationId": "listActresses",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "关键词搜索（匹配 name_romaji/name_kanji/name_kana）",
            "schema": { "type": "string" }
          },
          {
            "name": "page",
            "in": "query",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "name": "page_size",
            "in": "query",
            "schema": { "type": "integer", "default": 20, "maximum": 100 }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ActressListResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/actresses/{id}": {
      "get": {
        "summary": "演员详情",
        "description": "获取单个演员的信息",
        "operationId": "getActress",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "演员ID",
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ActressDetailResponse" }
              }
            }
          },
          "400": {
            "description": "id 不是正整数",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "演员不存在",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/actresses/{id}/videos": {
      "get": {
        "summary": "演员作品列表",
        "description": "获取某演员出演的所有视频",
        "operationId": "getActressVideos",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "演员ID",
            "schema": { "type": "integer" }
          },
          {
            "name": "page",
            "in": "query",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "name": "page_size",
            "in": "query",
            "schema": { "type": "integer", "default": 20, "maximum": 100 }
          },
          {
            "name": "service_code",
            "in": "query",
            "description": "服务代码筛选（digital/mono/rental/ebook），空值不过滤",
            "schema": { "type": "string" }
          },
          {
            "name": "year",
            "in": "query",
            "description": "发行年份筛选",
            "schema": { "type": "integer" }
          },
          {
            "name": "maker_id",
            "in": "query",
            "description": "厂商ID（精确匹配，优先于 maker_name）",
            "schema": { "type": "integer" }
          },
          {
            "name": "maker_name",
            "in": "query",
            "description": "厂商名称（模糊匹配）",
            "schema": { "type": "string" }
          },
          {
            "name": "category_id",
            "in": "query",
            "description": "分类ID（精确匹配，优先于 category_name）",
            "schema": { "type": "integer" }
          },
          {
            "name": "category_name",
            "in": "query",
            "description": "分类名称（模糊匹配）",
            "schema": { "type": "string" }
          },
          {
            "name": "sort_by",
            "in": "query",
            "description": "排序字段及方向，格式：field:dir（release_date/content_id/dvd_id/title_en/title_ja/runtime_mins）",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/VideoListResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/actresses/batch_videos": {
      "post": {
        "summary": "批量获取演员作品",
        "description": "根据演员ID列表批量获取作品，最多20个演员。返回 actress_id -> {total_count, videos} 映射",
        "operationId": "batchActressVideos",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["ids"],
                "properties": {
                  "ids": {
                    "type": "array",
                    "items": { "type": "integer" },
                    "maxItems": 20,
                    "description": "演员ID列表"
                  },
                  "page": {
                    "type": "integer",
                    "default": 1,
                    "description": "页码"
                  },
                  "page_size": {
                    "type": "integer",
                    "default": 20,
                    "maximum": 100,
                    "description": "每页数量"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功，返回 {actress_id: {total_count, videos}} 映射",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object",
                    "properties": {
                      "total_count": { "type": "integer" },
                      "videos": {
                        "type": "array",
                        "items": { "$ref": "#/components/schemas/VideoSummary" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/makers": {
      "get": {
        "summary": "厂商列表",
        "description": "获取制作厂商分页列表",
        "operationId": "listMakers",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "关键词搜索（匹配 name_en/name_ja）",
            "schema": { "type": "string" }
          },
          {
            "name": "page",
            "in": "query",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "name": "page_size",
            "in": "query",
            "schema": { "type": "integer", "default": 20, "maximum": 100 }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MakerListResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/labels": {
      "get": {
        "summary": "品牌列表",
        "description": "获取品牌分页列表",
        "operationId": "listLabels",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "关键词搜索（匹配 name_en/name_ja）",
            "schema": { "type": "string" }
          },
          {
            "name": "page",
            "in": "query",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "name": "page_size",
            "in": "query",
            "schema": { "type": "integer", "default": 20, "maximum": 100 }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LabelListResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/series": {
      "get": {
        "summary": "系列列表",
        "description": "获取影片系列分页列表",
        "operationId": "listSeries",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "关键词搜索（匹配 name_en/name_ja）",
            "schema": { "type": "string" }
          },
          {
            "name": "page",
            "in": "query",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "name": "page_size",
            "in": "query",
            "schema": { "type": "integer", "default": 20, "maximum": 100 }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SeriesListResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/categories": {
      "get": {
        "summary": "题材分类列表",
        "description": "获取题材分类分页列表",
        "operationId": "listCategories",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "关键词搜索（匹配 name_en/name_ja）",
            "schema": { "type": "string" }
          },
          {
            "name": "page",
            "in": "query",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "name": "page_size",
            "in": "query",
            "schema": { "type": "integer", "default": 20, "maximum": 100 }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CategoryListResponse" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "VideoListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/VideoSummary" }
          },
          "page": { "type": "integer" },
          "page_size": { "type": "integer" },
          "total_count": { "type": "integer" },
          "total_pages": { "type": "integer" }
        }
      },
      "VideoSummary": {
        "type": "object",
        "properties": {
          "content_id": { "type": "string" },
          "dvd_id": { "type": "string" },
          "title_en": { "type": "string" },
          "title_ja": { "type": "string" },
          "runtime_mins": { "type": "integer" },
          "release_date": { "type": "string", "format": "date" },
          "jacket_thumb_url": { "type": "string" },
          "site_id": { "type": "integer" },
          "service_code": { "type": "string" }
        }
      },
...
      "ActressListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ActressSummary" }
          },
          "page": { "type": "integer" },
          "page_size": { "type": "integer" },
          "total_count": { "type": "integer" },
          "total_pages": { "type": "integer" }
        }
      },
      "ActressSummary": {
        "type": "object",
        "description": "演员列表项，包含作品数量",
        "properties": {
          "id": { "type": "integer" },
          "name_romaji": { "type": "string" },
          "name_kanji": { "type": "string" },
          "name_kana": { "type": "string" },
          "image_url": { "type": "string" },
          "movie_count": { "type": "integer", "description": "作品数量" }
        }
      },
      "ActressDetailResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name_romaji": { "type": "string" },
          "name_kanji": { "type": "string" },
          "name_kana": { "type": "string" },
          "image_url": { "type": "string" }
        }
      },
...
      "Maker": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name_en": { "type": "string" },
          "name_ja": { "type": "string" }
        }
      },
      "MakerListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Maker" }
          },
          "page": { "type": "integer" },
          "page_size": { "type": "integer" },
          "total_count": { "type": "integer" },
          "total_pages": { "type": "integer" }
        }
      },
      "Label": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name_en": { "type": "string" },
          "name_ja": { "type": "string" }
        }
      },
      "LabelListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Label" }
          },
          "page": { "type": "integer" },
          "page_size": { "type": "integer" },
          "total_count": { "type": "integer" },
          "total_pages": { "type": "integer" }
        }
      },
      "Series": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name_en": { "type": "string" },
          "name_ja": { "type": "string" }
        }
      },
      "SeriesListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Series" }
          },
          "page": { "type": "integer" },
          "page_size": { "type": "integer" },
          "total_count": { "type": "integer" },
          "total_pages": { "type": "integer" }
        }
      },
      "Category": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name_en": { "type": "string" },
          "name_ja": { "type": "string" }
        }
      },
      "CategoryListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Category" }
          },
          "page": { "type": "integer" },
          "page_size": { "type": "integer" },
          "total_count": { "type": "integer" },
          "total_pages": { "type": "integer" }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      }
    }
  }
}